@@ -1,17 +1,21 @@
-<?php declare(strict_types = 1);
+<?php
+
+declare(strict_types=1);
 
 namespace Midnight\Permissions;
 
-use Interop\Container\ContainerInterface;
-use Interop\Container\Exception\ContainerException;
-use Interop\Container\Exception\NotFoundException;
 use Midnight\Permissions\Exception\InvalidPermissionException;
 use Midnight\Permissions\Exception\UnknownPermissionException;
+use Psr\Container\ContainerInterface;
+
+use function get_class;
+use function gettype;
+use function is_object;
+use function sprintf;
 
 final class PermissionService implements PermissionServiceInterface
 {
-    /** @var ContainerInterface */
-    private $container;
+    private ContainerInterface $container;
 
     public function __construct(ContainerInterface $container)
     {
@@ -23,8 +27,6 @@ final class PermissionService implements PermissionServiceInterface
      * @param mixed|null $resource
      * @throws UnknownPermissionException
      * @throws InvalidPermissionException
-     * @throws ContainerException
-     * @throws NotFoundException
      */
     public function isAllowed($user = null, string $permission, $resource = null): bool
     {
@@ -33,12 +35,14 @@ final class PermissionService implements PermissionServiceInterface
         }
         $permissionObject = $this->container->get($permission);
         if (!$permissionObject instanceof PermissionInterface) {
-            throw new InvalidPermissionException(sprintf(
-                '"%s" is not a permission. Expected an instance of %s, but got %s.',
-                $permission,
-                PermissionInterface::class,
-                is_object($permissionObject) ? get_class($permissionObject) : gettype($permissionObject)
-            ));
+            throw new InvalidPermissionException(
+                sprintf(
+                    '"%s" is not a permission. Expected an instance of %s, but got %s.',
+                    $permission,
+                    PermissionInterface::class,
+                    is_object($permissionObject) ? get_class($permissionObject) : gettype($permissionObject)
+                )
+            );
         }
         return $permissionObject->isAllowed($user, $resource);
     }
