@@ -1,4 +1,6 @@
-<?php declare(strict_types = 1);
+<?php
+
+declare(strict_types=1);
 
 namespace Midnight\Permissions;
 
@@ -7,6 +9,7 @@ interface PermissionServiceInterface
     /**
      * @param mixed|null $user
      * @param mixed|null $resource
+     * @return bool
      */
     public function isAllowed($user = null, string $permission, $resource = null);
 }
